+++ packages/dapp/src/components/VCTable/GlobalFilter/index.tsx
@@ -1,4 +1,4 @@
-import { QueryVCsRequestResult } from '@blockchain-lab-um/masca-connector';
+import { QueryCredentialsRequestResult } from '@blockchain-lab-um/masca-connector';
 import { MagnifyingGlassIcon } from '@heroicons/react/20/solid';
 import { useTranslations } from 'next-intl';
 import { shallow } from 'zustand/shallow';
@@ -7,7 +7,7 @@ import { useTableStore } from '@/stores';
 
 type GlobalFilterProps = {
   isConnected: boolean;
-  vcs: QueryVCsRequestResult[];
+  vcs: QueryCredentialsRequestResult[];
 };
 
 const GlobalFilter = ({ isConnected, vcs }: GlobalFilterProps) => {