+++ packages/dapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blockchain-lab-um/dapp
 
+## 0.1.1-beta.1
+
+### Patch Changes
+
+- [#355](https://github.com/blockchain-lab-um/masca/pull/355) [`527bd50`](https://github.com/blockchain-lab-um/masca/commit/527bd50b8d79a80f5637103975dbaebaf102a1e6) Thanks [@andyv09](https://github.com/andyv09)! - Improved popups on Masca and implemented friendly dApps functionality
+
+- [#369](https://github.com/blockchain-lab-um/masca/pull/369) [`cf5b2a1`](https://github.com/blockchain-lab-um/masca/commit/cf5b2a13b997e819733e589c4f51be808d4d18b1) Thanks [@andyv09](https://github.com/andyv09)! - Change RPC method names and internal types
+
+- Updated dependencies [[`527bd50`](https://github.com/blockchain-lab-um/masca/commit/527bd50b8d79a80f5637103975dbaebaf102a1e6), [`cf5b2a1`](https://github.com/blockchain-lab-um/masca/commit/cf5b2a13b997e819733e589c4f51be808d4d18b1), [`881b59b`](https://github.com/blockchain-lab-um/masca/commit/881b59b7cb59ae89662294172ca9b206b3ea1d57)]:
+  - @blockchain-lab-um/masca-connector@0.4.2-beta.1
+  - @blockchain-lab-um/oidc-types@0.0.2-beta.0
+
 ## 0.1.1-beta.0
 
 ### Patch Changes